% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{CheckMaxDiffDesign}
\alias{CheckMaxDiffDesign}
\title{\code{CheckMaxDiffDesign}}
\usage{
CheckMaxDiffDesign(design)
}
\arguments{
\item{design}{A \code{\link{matrix}}, where each row represents a question or task, and each column
shows the alternatives to be shown.}
}
\description{
Produces summary statistics for a MaxDiff design.
}
