% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probabilities.R
\name{dBestWorst}
\alias{dBestWorst}
\title{\code{dBestWorst}}
\usage{
dBestWorst(e.u)
}
\arguments{
\item{e.u}{A vector of \code{exp(utilities)}, where items are shown in the order chosen, from 'best' to 'worst'.}
}
\description{
Computes the probabilities that a pair of items are chosen as 'best' and 'worst' from a set.
}
