% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probabilities.R
\name{dExplodedLogit}
\alias{dExplodedLogit}
\title{\code{dExplodedLogit}}
\usage{
dExplodedLogit(e.u)
}
\arguments{
\item{e.u}{A vector of exp(utilities), ordered from most 'preferred to leas'best' to 'worst'.}
}
\description{
Computes the probability of an ordering, where the items are ordered from 'best' to 'worst'.
}
