% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max-difflogit.R
\name{dMaxDiff}
\alias{dMaxDiff}
\title{\code{dMaxDiff}}
\usage{
dMaxDiff(b, x)
}
\arguments{
\item{b}{A vector of parameter estimates.}

\item{x}{The experimental design for a respondent  (a \code{\link{list}}).}
}
\description{
The density of a respondent's max-diff data.
}
