% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max-diff.R
\name{Memberships}
\alias{Memberships}
\title{\code{Memberships}}
\usage{
Memberships(object)
}
\arguments{
\item{object}{A \code{FitMaxDiff} object.}
}
\description{
Allocates respondents to classes based on the maximum posterior probabilty.
}
