% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{multipleVersionDesign}
\alias{multipleVersionDesign}
\title{\code{multipleVersionDesign}}
\usage{
multipleVersionDesign(original, n.versions)
}
\arguments{
\item{original}{The experiment design that is randomized to individual-level vesions.}

\item{n.versions}{The number of versions of the experimental design (defaults to 1). Subsequent versions are obtained by permutting the columns of the binary design.}
}
\description{
\code{multipleVersionDesign}
}
