% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probabilities.R
\name{Permutations}
\alias{Permutations}
\title{\code{permutations}}
\usage{
Permutations(x, prefix = c())
}
\arguments{
\item{x}{A vector.}

\item{prefix}{The empty case; used for recursion - don't play with this.}
}
\description{
Computes all possible permitatuions of \code{x}.
}
\references{
http://stackoverflow.com/a/29023189/1547926
}
