% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max-diff.R
\name{print.FitMaxDiff}
\alias{print.FitMaxDiff}
\title{\code{print.CorrespondenceAnalysis}}
\usage{
\method{print}{FitMaxDiff}(x, ...)
}
\arguments{
\item{x}{FitMaxDiff object.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
\code{print.CorrespondenceAnalysis}
}
