% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{CheckMaxDiffDesign}
\alias{CheckMaxDiffDesign}
\title{\code{CheckMaxDiffDesign}}
\usage{
CheckMaxDiffDesign(design)
}
\arguments{
\item{design}{The design.}
}
\description{
Produces summary statistics for a max-diff design.
}
