% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max-diff.R
\name{FitMaxDiff}
\alias{FitMaxDiff}
\title{\code{FitMaxDiff}}
\usage{
FitMaxDiff(
  design,
  version = NULL,
  best,
  worst,
  alternative.names,
  n.classes = 1,
  subset = NULL,
  weights = NULL,
  characteristics = NULL,
  seed = 123,
  initial.parameters = NULL,
  trace = 0,
  sub.model.outputs = FALSE,
  lc = TRUE,
  output = "Probabilities",
  tasks.left.out = 0,
  is.mixture.of.normals = FALSE,
  algorithm = "Default",
  normal.covariance = "Full",
  pool.variance = FALSE,
  lc.tolerance = 1e-04,
  n.draws = 100,
  is.tricked = FALSE,
  hb.iterations = 100,
  hb.chains = 1,
  hb.max.tree.depth = 10,
  hb.adapt.delta = 0.8,
  use.krinsky.robb = FALSE,
  show.mnl.details = FALSE
)
}
\arguments{
\item{design}{A \code{data.frame}, where the first variable is called 'Version', the second is called 'Task' or 'Question',
and the remaining variables contain the alternatives shown in each task.}

\item{version}{A vector of integers showing the version of the design shown to each respondent.}

\item{best}{A data frame of factors or a matrix of integers showing the choices made by each respondent on each of the questions. One column
for each task. The integers need to correspond to the \code{design} vector of integers showing the version of
the design shown to each respondent. Coerced to a matrix if a \code{data.frame}.}

\item{worst}{As with 'best', except denoting worst..}

\item{alternative.names}{A \code{character} vector of the alternative names. Where \code{best}
and  \code{worst} are factors or characters, these names must match them.}

\item{n.classes}{The number of latent classes.}

\item{subset}{An optional vector specifying a subset of observations to be
used in the fitting process.}

\item{weights}{An optional vector of sampling or frequency weights.}

\item{characteristics}{Data frame of characteristics on which to run varying coefficients by latent class boosting.}

\item{seed}{Seed for initial random class assignments.}

\item{initial.parameters}{Specify initial parameters intead of starting at random.}

\item{trace}{Non-negative integer indicating the detail of outputs provided when fitting models: 0 indicates
no outputs, and 6 is the most detailed outputs.}

\item{sub.model.outputs}{If TRUE, prints diagnostics on interim models.}

\item{lc}{Whether to run latent class step at the end if characteristics are supplied.}

\item{output}{Output type. Can be "Probabilities" or "Classes".}

\item{tasks.left.out}{Number of questions to leave out for cross-validation.}

\item{is.mixture.of.normals}{Whether to model with mixture of normals instead of LCA.}

\item{algorithm}{If "HB", Hierarchical Bayes with a MVN prior is used and the other parameters are ignored.}

\item{normal.covariance}{The form of the covariance matrix for mixture of normals.
Can be 'Full, 'Spherical', 'Diagonal'.}

\item{pool.variance}{Whether to pool parameter covariances between classes in mixture of normals.}

\item{lc.tolerance}{The tolerance used for defining convergence in latent class analysis.}

\item{n.draws}{The number of draws when fitting mixture of normals.}

\item{is.tricked}{Whether to use tricked logit instead of rank-ordered logit with ties.}

\item{hb.iterations}{The number of iterations in Hierarchical Bayes.}

\item{hb.chains}{The number of chains in Hierarchical Bayes.}

\item{hb.max.tree.depth}{http://mc-stan.org/misc/warnings.html#maximum-treedepth-exceeded}

\item{hb.adapt.delta}{http://mc-stan.org/misc/warnings.html#divergent-transitions-after-warmup}

\item{show.mnl.details}{Logical. If TRUE, prints MNL model coefficients, Krinsky-Robb standard errors, and preference shares.}
}
\description{
Fits a latent class rank-ordered logit model with ties to a max-diff experiment.
}
