% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max-difflogit.R
\name{FitMaxDiff}
\alias{FitMaxDiff}
\title{\code{FitMaxDiff}}
\usage{
FitMaxDiff(design, version, best, worst)
}
\arguments{
\item{design}{A \code{data.frame}, where the first variable is called 'Version', the second is called 'Task',
and the remaining variables contain the alternatives shown in each tas.}

\item{version}{A vector of integers showing the version of the design shown to each respondent.}

\item{best}{A matrix of integers showing the choices made by each respondent on each of the tasks. One column
for each task. The integers need to corresponde to the \code{design} vector of integers showing the version of
the design shown to each respondent. Coerced to a matrix if a \code{data.frame}.}

\item{worst}{A matrix of integers showing the choice of 'worst'.}
}
\description{
Fits a rank-ordered logit model with ties to a max-diff experiment.
}
