% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max-difflogit.R
\name{FitMaxDiff}
\alias{FitMaxDiff}
\title{\code{FitMaxDiff}}
\usage{
FitMaxDiff(design, version = NULL, best, worst, alternative.names,
  n.classes = 1, subset = NULL, weights = NULL, characteristics = NULL,
  seed = 123, initial.parameters = NULL, trace = 0,
  sub.model.outputs = FALSE, lc = TRUE, output = "Probabilities",
  tasks.left.out = 0)
}
\arguments{
\item{design}{A \code{data.frame}, where the first variable is called 'Version', the second is called 'Task' or 'Question',
and the remaining variables contain the alternatives shown in each task.}

\item{version}{A vector of integers showing the version of the design shown to each respondent.}

\item{best}{A data frame of factors or a matrix of integers showing the choices made by each respondent on each of the tasks. One column
for each task. The integers need to correspond to the \code{design} vector of integers showing the version of
the design shown to each respondent. Coerced to a matrix if a \code{data.frame}.}

\item{worst}{As with 'best', except denoting worst..}

\item{alternative.names}{A \code{character} vector of the alternative names. Where \code{best}
and  \code{worst} are factors or characters, these names must match them.}

\item{n.classes}{The number of latent classes.}

\item{subset}{An optional vector specifying a subset of observations to be
used in the fitting process.}

\item{weights}{An optional vector of sampling or frequency weights.}

\item{characteristics}{Data frame of characteristics on which to run varying coefficients by latent class boosting.}

\item{seed}{Seed for initial random class assignments.}

\item{initial.parameters}{Specify initial parameters intead of starting at random.}

\item{trace}{Non-negative integer indicating the detail of outputs provided when fitting models: 0 indicates}

\item{sub.model.outputs}{If TRUE, prints diagnostics on interrum models.}

\item{lc}{Whether to run latent class step at the end if characteristics are supplied.
no outputs, and 6 is the most detailed outputs.}

\item{output}{Output type. Can be "Probabilities" or "Classes".}

\item{tasks.left.out}{Number of tasks to leave out for cross-validation.}
}
\description{
Fits a latent class rank-ordered logit model with ties to a max-diff experiment.
}
