% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max-difflogit.R
\name{FitMaxDiff}
\alias{FitMaxDiff}
\title{\code{FitMaxDiff}}
\usage{
FitMaxDiff(design, version, best, worst, names, subset = NULL,
  weights = NULL, trace = 0)
}
\arguments{
\item{design}{A \code{data.frame}, where the first variable is called 'Version', the second is called 'Task',
and the remaining variables contain the alternatives shown in each tas.}

\item{version}{A vector of integers showing the version of the design shown to each respondent.}

\item{best}{A matrix of integers showing the choices made by each respondent on each of the tasks. One column
for each task. The integers need to corresponde to the \code{design} vector of integers showing the version of
the design shown to each respondent. Coerced to a matrix if a \code{data.frame}.}

\item{worst}{A matrix of integers showing the choice of 'worst'.}

\item{names}{A character vector names of the alternatives. If only a single element is supplied, it is split by commas.}

\item{subset}{An optional vector specifying a subset of observations to be
used in the fitting process.}

\item{weights}{An optional vector of sampling or frequency weights.}

\item{trace}{Non-negative integer indicating the detail of outputs provided during estimation: 0 indicates
no outputs, and 6 is the most detailed outputs.}
}
\description{
Fits a rank-ordered logit model with ties to a max-diff experiment.
}
