% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max-difflogit.R
\name{logLikelihoodMaxDiff}
\alias{logLikelihoodMaxDiff}
\title{\code{dMaxDiff}}
\usage{
logLikelihoodMaxDiff(b, X)
}
\arguments{
\item{b}{A vector of parameter estimates.}

\item{X}{The experimental design for a sample (a \code{\link{list}})}
}
\description{
The log-likelihood for a max-diff experiment.
}
