% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.R
\name{logLikelihoodMaxDiff}
\alias{logLikelihoodMaxDiff}
\title{\code{dMaxDiff}}
\usage{
logLikelihoodMaxDiff(b, X, boost, weights)
}
\arguments{
\item{b}{A vector of parameter estimates.}

\item{X}{The experimental design for a sample (a \code{\link{list}})}

\item{boost}{Boosting constants.}

\item{weights}{An optional vector of sampling or frequency weights.}
}
\description{
The log-likelihood for a MaxDiff experiment.
}
