% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max-diff.R
\name{RespondentParameters}
\alias{RespondentParameters}
\title{\code{RespondentParameters}}
\usage{
RespondentParameters(object)
}
\arguments{
\item{object}{A \code{FitMaxDiff} object.}
}
\description{
The parameters for each respondent.
}
